extract common parts of job roles selection pages into a directive for reuse
