+++ frontend/src/app/core/model/establishment.model.ts
@@ -58,12 +58,14 @@ export interface Starter {
   jobId: number;
   title?: string;
   total: number;
+  other?: string;
 }
 
 export interface Leaver {
   jobId: number;
   title?: string;
   total: number;
+  other?: string;
 }
 
 export interface Category {
