+++ frontend/src/app/features/workplace/vacancies-job-roles-selection/vacancies-job-roles-selection.component.ts
@@ -1,128 +1,38 @@
-import { Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
-import { UntypedFormBuilder } from '@angular/forms';
-import { ActivatedRoute, Router } from '@angular/router';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Vacancy } from '@core/model/establishment.model';
-import { Job, JobGroup } from '@core/model/job.model';
-import { BackService } from '@core/services/back.service';
-import { ErrorSummaryService } from '@core/services/error-summary.service';
-import { EstablishmentService } from '@core/services/establishment.service';
-import { JobService } from '@core/services/job.service';
-import { Question } from '@features/workplace/question/question.component';
-import { AccordionGroupComponent } from '@shared/components/accordions/generic-accordion/accordion-group/accordion-group.component';
-import { CustomValidators } from '@shared/validators/custom-form-validators';
+import { JobRoleSelectionDirective } from '../vacancies-and-turnover/job-role-selection.directive';
 
 @Component({
   selector: 'app-vacancies-job-roles-selection',
-  templateUrl: './vacancies-job-roles-selection.component.html',
+  templateUrl: '../vacancies-and-turnover/job-role-selection.html',
 })
-export class VacanciesJobRolesSelectionComponent extends Question implements OnInit, OnDestroy {
-  @ViewChild('accordion') accordion: AccordionGroupComponent;
-  public section = 'Vacancies and turnover';
+export class VacanciesJobRolesSelectionComponent extends JobRoleSelectionDirective implements OnInit, OnDestroy {
   public errorMessageOnEmptyInput = 'Select job roles for all your current staff vacancies';
-  public jobIdOfOtherCareProvidingRole = 20;
-  public jobGroupsToOpenAtStart: string[] = [];
-  public jobGroups: JobGroup[] = [];
-
+  public heading = 'Select job roles for all your current staff vacancies';
   protected localStorageKey = 'updated-vacancies';
+  protected prefillData: Vacancy[] = [];
 
-  private jobsAvailable: Job[] = [];
-  private vacancies: Vacancy[] = [];
-  private prefilledJobIds: number[] = [];
-
-  constructor(
-    protected formBuilder: UntypedFormBuilder,
-    protected router: Router,
-    protected backService: BackService,
-    protected errorSummaryService: ErrorSummaryService,
-    protected establishmentService: EstablishmentService,
-    private route: ActivatedRoute,
-  ) {
-    super(formBuilder, router, backService, errorSummaryService, establishmentService);
-  }
-
-  protected init() {
-    this.getJobs();
-    this.setupForm();
-    this.prefill();
-
+  protected setupRoutes(): void {
     this.nextRoute = ['/workplace', `${this.establishment.uid}`, 'vacancies-number'];
     this.previousRoute = ['/workplace', `${this.establishment.uid}`, 'vacancies'];
   }
 
-  private getJobs(): void {
-    this.jobsAvailable = this.route.snapshot.data.jobs;
-    this.jobGroups = JobService.sortJobsByJobGroup(this.jobsAvailable);
-  }
-
-  private setupForm(): void {
-    this.form = this.formBuilder.group({
-      selectedJobRoles: [[], CustomValidators.validateArrayNotEmpty()],
-      otherCareProvidingRoleName: ['', null],
-    });
-  }
-
-  private prefill(): void {
-    this.vacancies = this.getPrefillData() ?? [];
-    if (!this.vacancies?.length) {
-      return;
-    }
-
-    this.prefilledJobIds = this.vacancies.map((vacancy) => Number(vacancy.jobId));
-    this.jobGroupsToOpenAtStart = this.jobGroups
-      .filter((group) => group.items.some((job) => this.prefilledJobIds.includes(job.id)))
-      .map((group) => group.title);
-
-    const otherCareProvidingRole = this.vacancies.find(
-      (vacancy) => vacancy.jobId === this.jobIdOfOtherCareProvidingRole,
-    );
-    this.form.patchValue({
-      selectedJobRoles: this.prefilledJobIds,
-      otherCareProvidingRoleName: otherCareProvidingRole?.other ?? null,
-    });
-  }
-
-  private getPrefillData() {
+  protected getPrefillData() {
     const previousData = this.loadFromLocal();
     if (previousData?.establishmentUid === this.establishment.uid && Array.isArray(previousData?.vacancies)) {
-      return previousData.vacancies;
+      this.prefillData = previousData.vacancies;
+    } else if (Array.isArray(this.establishment.vacancies)) {
+      this.prefillData = this.establishment.vacancies;
     }
-
-    if (Array.isArray(this.establishment.vacancies)) {
-      return this.establishment.vacancies;
-    }
-  }
-
-  public onCheckboxClick(target: HTMLInputElement) {
-    const jobId = Number(target.value);
-    const selectedJobRoles = this.form.get('selectedJobRoles');
-    const currentSelectedIds: number[] = selectedJobRoles.value;
-
-    if (currentSelectedIds.includes(jobId)) {
-      const updatedSelection = currentSelectedIds.filter((id) => id !== jobId);
-      selectedJobRoles.setValue(updatedSelection);
-    } else {
-      selectedJobRoles.setValue([...currentSelectedIds, jobId]);
-    }
-  }
-
-  public onSubmit(): void {
-    if (this.form.invalid) {
-      this.accordion.showAll();
-    }
-    super.onSubmit();
   }
 
   protected onSuccess(): void {
-    this.storeCurrentChanges();
-  }
-
-  protected storeCurrentChanges(): void {
     const selectedJobIds: number[] = this.form.get('selectedJobRoles').value;
     const otherCareProvidingRoleName: string = this.form.get('otherCareProvidingRoleName').value;
 
     const updatedVacancies: Vacancy[] = selectedJobIds.map((jobId) => {
       const job = this.jobsAvailable.find((job) => job.id === jobId);
-      const vacancyCount = this.vacancies.find((vacancy) => vacancy.jobId === jobId)?.total ?? null;
+      const vacancyCount = this.prefillData.find((vacancy) => vacancy.jobId === jobId)?.total ?? null;
       if (job.id === this.jobIdOfOtherCareProvidingRole && otherCareProvidingRoleName) {
         return { jobId, title: job.title, total: vacancyCount, other: otherCareProvidingRoleName };
       }
@@ -131,36 +41,5 @@ export class VacanciesJobRolesSelectionComponent extends Question implements OnI
     });
     const dataToStore = { establishmentUid: this.establishment.uid, vacancies: updatedVacancies };
     this.saveToLocal(dataToStore);
-  }
-
-  protected saveToLocal(dataToStore) {
-    localStorage.setItem(this.localStorageKey, JSON.stringify(dataToStore));
-  }
-
-  protected loadFromLocal() {
-    try {
-      return JSON.parse(localStorage.getItem(this.localStorageKey));
-    } catch (err) {
-      return null;
-    }
-  }
-
-  protected setupFormErrorsMap(): void {
-    this.formErrorsMap = [
-      {
-        item: 'selectedJobRoles',
-        type: [
-          {
-            name: 'selectedNone',
-            message: this.errorMessageOnEmptyInput,
-          },
-        ],
-      },
-    ];
-  }
-
-  protected generateUpdateProps() {
-    // suppress the default action of making calls to backend
-    return null;
   }
 }
