+++ frontend/src/app/shared/components/accordions/generic-accordion/accordion-group/accordion-group.component.spec.ts
@@ -3,7 +3,7 @@ import { SharedModule } from '@shared/shared.module';
 import { render, within } from '@testing-library/angular';
 import userEvent from '@testing-library/user-event';
 
-fdescribe('AccordionGroupComponent', () => {
+describe('AccordionGroupComponent', () => {
   const setup = async (override: any = {}) => {
     const componentProps = {
       contentName: override.contentName,
